4-final-project/requirements.txt
Django==2.0
django-environ==0.4.4
djangorestframework==3.7.3
djangorestframework-gis==0.12.0
markdown
django-filter
psycopg2
